adjustments to make sure Helm charts are public, images private
